requirements.txt
PyQt6
PyMuPDF
Pillow
google-generativeai
python-dotenv